chore: more restrictive semver version ranges (#756)
We currently validate our version range by sending it to
`semver.validRange`. However, this function is rather permissive and
does not flag ranges that are probably errors, but still valid. For
example, `2.0.0 <= 2.1000.0` is valid semver. It just evaluates to
`2.0.0` _only_.

I've decided that the right thing to do here is to require a more
restrictive semver range. One where if you include a greater than symbol
(i.e. `^, >=, >`) you must also include a less than symbol (i.e. `<,
<='). This means we no longer are allowed to write `<= 2.1000.0` for
example, but we haven't done that yet in our existing notices and it
makes sense to require both a lower and upper bound on the ranges we
provide.

---------

Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>